requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

gns3_cx_Freeze==4.3.6
pygame==2.1.2
